Move triton runtime code to subdirectory
