+++ merlin/systems/dag/runtimes/triton/runtime.py
@@ -13,9 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+import importlib.resources
 import os
 import pathlib
-from shutil import copyfile, copytree
+from shutil import copyfile
 from typing import List, Tuple
 
 from merlin.dag import postorder_iter_nodes
@@ -27,12 +28,11 @@ import tritonclient.grpc.model_config_pb2 as model_config  # noqa
 from google.protobuf import text_format  # noqa
 
 from merlin.core.protocols import Transformable  # noqa
-from merlin.dag import ColumnSelector, Graph  # noqa
-from merlin.schema import Schema  # noqa
+from merlin.dag import Graph  # noqa
 from merlin.systems.dag.ops import compute_dims  # noqa
-from merlin.systems.dag.ops.compat import pb_utils  # noqa
 from merlin.systems.dag.ops.operator import add_model_param  # noqa
 from merlin.systems.dag.runtimes import Runtime  # noqa
+from merlin.systems.dag.runtimes.triton.ops.tensorflow import PredictTensorflowTriton  # noqa
 
 tensorflow = None
 try:
@@ -40,164 +40,11 @@ try:
 except ImportError:
     ...
 
+TRITON_OP_TABLE = {}
+if tensorflow:
+    from merlin.systems.dag.ops.tensorflow import PredictTensorflow
 
-class TritonOperator:
-    def __init__(self, base_op):
-        self.op = base_op
-        self.input_schema = self.op.input_schema
-        self.output_schema = self.op.output_schema
-
-    @property
-    def export_name(self):
-        """
-        Provides a clear common english identifier for this operator.
-
-        Returns
-        -------
-        String
-            Name of the current class as spelled in module.
-        """
-        return self.__class__.__name__.lower()
-
-    @property
-    def exportable_backends(self) -> List[str]:
-        """Returns list of supported backends.
-
-        Returns
-        -------
-        List[str]
-            List of supported backends
-        """
-        return ["ensemble"]
-
-
-class PredictTensorflowTriton(TritonOperator):
-    """TensorFlow Model Prediction Operator for running inside Triton."""
-
-    def __init__(self, base_op):
-        super().__init__(base_op)
-
-        self.path = self.op.path
-        self.model = self.op.model
-        self.scalar_shape = self.op.scalar_shape
-
-    def transform(self, col_selector: ColumnSelector, transformable: Transformable):
-        """Run transform of operator callling TensorFlow model with a Triton InferenceRequest.
-
-        Returns
-        -------
-        Transformable
-            TensorFlow Model Outputs
-        """
-        # TODO: Validate that the inputs match the schema
-        # TODO: Should we coerce the dtypes to match the schema here?
-        input_tensors = []
-        for col_name in self.input_schema.column_schemas.keys():
-            input_tensors.append(pb_utils.Tensor(col_name, transformable[col_name]))
-
-        inference_request = pb_utils.InferenceRequest(
-            model_name=self.tf_model_name,
-            requested_output_names=self.output_schema.column_names,
-            inputs=input_tensors,
-        )
-        inference_response = inference_request.exec()
-
-        # TODO: Validate that the outputs match the schema
-        outputs_dict = {}
-        for out_col_name in self.output_schema.column_schemas.keys():
-            output_val = pb_utils.get_output_tensor_by_name(
-                inference_response, out_col_name
-            ).as_numpy()
-            outputs_dict[out_col_name] = output_val
-
-        return type(transformable)(outputs_dict)
-
-    def export(
-        self,
-        path: str,
-        input_schema: Schema,
-        output_schema: Schema,
-        params: dict = None,
-        node_id: int = None,
-        version: int = 1,
-        backend: str = "ensemble",
-    ):
-        """Create a directory inside supplied path based on our export name"""
-        # Export Triton TF back-end directory and config etc
-        export_name = self.__class__.__name__.lower()
-        node_name = f"{node_id}_{export_name}" if node_id is not None else export_name
-
-        node_export_path = pathlib.Path(path) / node_name
-        node_export_path.mkdir(exist_ok=True)
-
-        tf_model_path = pathlib.Path(node_export_path) / str(version) / "model.savedmodel"
-
-        if self.path:
-            copytree(
-                str(self.path),
-                tf_model_path,
-                dirs_exist_ok=True,
-            )
-        else:
-            self.model.save(tf_model_path, include_optimizer=False)
-
-        self.set_tf_model_name(node_name)
-        backend_model_config = self._export_model_config(node_name, node_export_path)
-        return backend_model_config
-
-    def _export_model_config(self, name, output_path):
-        """Exports a TensorFlow model for serving with Triton
-
-        Parameters
-        ----------
-        model:
-            The tensorflow model that should be served
-        name:
-            The name of the triton model to export
-        output_path:
-            The path to write the exported model to
-        """
-        config = model_config.ModelConfig(
-            name=name, backend="tensorflow", platform="tensorflow_savedmodel"
-        )
-
-        config.parameters["TF_GRAPH_TAG"].string_value = "serve"
-        config.parameters["TF_SIGNATURE_DEF"].string_value = "serving_default"
-
-        for _, col_schema in self.input_schema.column_schemas.items():
-            add_model_param(
-                config.input,
-                model_config.ModelInput,
-                col_schema,
-                compute_dims(col_schema, self.scalar_shape),
-            )
-
-        for _, col_schema in self.output_schema.column_schemas.items():
-            add_model_param(
-                config.output,
-                model_config.ModelOutput,
-                col_schema,
-                compute_dims(col_schema, self.scalar_shape),
-            )
-
-        with open(os.path.join(output_path, "config.pbtxt"), "w", encoding="utf-8") as o:
-            text_format.PrintMessage(config, o)
-        return config
-
-    @property
-    def tf_model_name(self):
-        return self._tf_model_name
-
-    def set_tf_model_name(self, tf_model_name: str):
-        """
-        Set the name of the Triton model to use
-
-        Parameters
-        ----------
-        tf_model_name : str
-            Triton model directory name
-        """
-        self._tf_model_name = tf_model_name
+    TRITON_OP_TABLE[PredictTensorflow] = PredictTensorflowTriton
 
 
 class TritonEnsembleRuntime(Runtime):
@@ -205,11 +52,7 @@ class TritonEnsembleRuntime(Runtime):
 
     def __init__(self):
         super().__init__()
-        self.op_table = {}
-        if tensorflow:
-            from merlin.systems.dag.ops.tensorflow import PredictTensorflow
-
-            self.op_table[PredictTensorflow] = PredictTensorflowTriton
+        self.op_table = TRITON_OP_TABLE
 
     def transform(self, graph: Graph, transformable: Transformable):
         raise NotImplementedError("Transform handled by Triton")
@@ -368,11 +211,7 @@ class TritonExecutorRuntime(Runtime):
 
     def __init__(self):
         super().__init__()
-        self.op_table = {}
-        if tensorflow:
-            from merlin.systems.dag.ops.tensorflow import PredictTensorflow
-
-            self.op_table[PredictTensorflow] = PredictTensorflowTriton
+        self.op_table = TRITON_OP_TABLE
 
     def export(
         self, ensemble, path: str, version: int = 1, name: str = None
@@ -484,17 +323,13 @@ class TritonExecutorRuntime(Runtime):
 
         os.makedirs(node_export_path, exist_ok=True)
         os.makedirs(os.path.join(node_export_path, str(version)), exist_ok=True)
-        copyfile(
-            os.path.join(
-                os.path.dirname(__file__),
-                "..",
-                "..",
-                "triton",
-                "models",
-                "executor_model.py",
-            ),
-            os.path.join(node_export_path, str(version), "model.py"),
-        )
+        with importlib.resources.path(
+            "merlin.systems.triton.models", "executor_model.py"
+        ) as executor_model:
+            copyfile(
+                executor_model,
+                os.path.join(node_export_path, str(version), "model.py"),
+            )
 
         ensemble.save(os.path.join(node_export_path, str(version), "ensemble"))
