+++ tests/unit/systems/ops/tf/test_op.py
@@ -31,6 +31,7 @@ from google.protobuf import text_format  # noqa
 from tritonclient.grpc import model_config_pb2 as model_config  # noqa
 
 tf_op = pytest.importorskip("merlin.systems.dag.ops.tensorflow")
+tf_triton_op = pytest.importorskip("merlin.systems.dag.runtimes.triton.ops.tensorflow")
 
 tf = pytest.importorskip("tensorflow")
 
@@ -55,7 +56,8 @@ def test_tf_op_exports_own_config(tmpdir):
     output_schema = Schema([ColumnSchema("output", dtype=np.float32)])
 
     # Triton
-    triton_op = tf_op.PredictTensorflowTriton(model)
+    tf_model_op = tf_op.PredictTensorflow(model)
+    triton_op = tf_triton_op.PredictTensorflowTriton(tf_model_op)
     triton_op.export(tmpdir, input_schema, output_schema)
 
     # Export creates directory
